+++ src/game/game.hpp
@@ -323,6 +323,7 @@ public:
 	void playerCyclopediaCharacterInfo(std::shared_ptr<Player> player, uint32_t characterID, CyclopediaCharacterInfoType_t characterInfoType, uint16_t entriesPerPage, uint16_t page);
 
 	void playerHighscores(std::shared_ptr<Player> player, HighscoreType_t type, uint8_t category, uint32_t vocation, const std::string &worldName, uint16_t page, uint8_t entriesPerPage);
+	static std::string getSkillNameById(uint8_t &skill);
 
 	void updatePlayerSaleItems(uint32_t playerId);
 
@@ -738,11 +739,16 @@ public:
 	Badge getBadgeById(uint8_t id);
 	Badge getBadgeByName(const std::string &name);
 
+	std::unordered_set<Title> getTitles();
+	Title getTitleById(uint8_t id);
+	Title getTitleByName(const std::string &name);
+
 private:
 	std::map<uint16_t, Achievement> m_achievements;
 	std::map<std::string, uint16_t> m_achievementsNameToId;
 
 	std::unordered_set<Badge> m_badges;
+	std::unordered_set<Title> m_titles;
 
 	std::vector<HighscoreCategory> m_highscoreCategories;
 	std::unordered_map<uint8_t, std::string> m_highscoreCategoriesNames;