+++ markdowns/CHANGELOG.md
@@ -11,6 +11,7 @@
   - Badge system ([elsongabriel](https://github.com/elsongabriel))
   - Screenshots configurable in config.lua enableScreenshots ([Tryller](https://github.com/jprzimba)).
   - Augments system ([phacUFPE](https://github.com/phacUFPE)).
+  - Add Cyclopedia character titles ([Tryller](https://github.com/jprzimba)).
 
 ### Bug Fixes
   - Fixed chain system. Now works properly for Mages and Paladins ([Tryller](https://github.com/jprzimba)).